Fix clickEvent being put on wrong element for tables
Currently, code for adding the clickEvent when used with a button
for a table results in the clickEvent being put on the
nextElementSibling of the table. This could be null or an unrelated
element.

This fix moves the button creation to before the registration of the
clickEvent, ensuring that the button is always the nextElementSibling.

fixes https://github.com/tomik23/show-more/issues/52